Refresh when the app enters the foreground, plus cache invalidation
The app should now refresh whenever it is brought into the foreground,
not just when loading the view from scratch.

In addition, I set the cache invalidation for 10 minutes. Some
experimentation will determine whether this is best amount of time. The
cache is always invalidated whenever the view loads from scratch. There
should also be a refresh control somewhere that invalidates the cache
and refreshes.