Bug Fixes
Fixed the helper method returning to % characters.

Also fixed the high and low temps for the current day. Forgot that I
have to get those from the first daily data point, not the currently
data point.

We need to check the helper methods to make sure that they return nil
if a nil string is passed in.